Opus: Fix integer bug in header parsing. Fixes issue #581.
Fix errors in parsing an OggOpus header packet where aliased pointers of
different type widths are used with psf_binheader_readf(), resulting in
incorrect data or endian issues. Telling psf_binheader_readf() to read
an integer of fixed width, but then passing a pointer to an integer
of a different width is a bug.