id3: Fix skip of large ID3v2 headers
Fix a bug in id3.c where large ID3v2 headers would fail to seek to the
embedded file. ID3v2 tags are skipped over, treating the wrapped file as
an embedded file. id3_skip() sets the embedded file offset to be just
after the ID3v2 tag ends, then seeks the psf_binheader() to the
fileoffset.

psf_binheader_read() accomplishes seeks by either reading and buffering
bytes until the offset is reached, or if the buffer would be too large,
seeking the underlying file instead using psf_fseek(). psf_fread does
not take the embedded file offset into account, while psf_fseek() does.

In the case of ID3 tags larger than the binheader buffer max, this has
the effect of seeking to twice the id3v2 tag length, as
psf_fseek(fileoffset, SEEK_CUR) adds the fileoffset to the offset
argument again. This lands in the middle of a stream which then cannot
be identified.

To resolve the issue, seek the binheader BEFORE setting the embedded
file offset.

An as-yet unresolved issue is that as skipping forward by seeks is not
supported on pipes, id3_skip() of large ID3v2 tags does not work on
pipes.